📝 Update TODO with what's going on
Now much has really been achieved here in the past ten years,
as I've been working on other projects.

Drop the idea of a C++
version of the project, but there is a potential for a Java version
of the project.

Remove obsolete URLs.  This was because bitbucket really dropped the ball
when they removed Mercurial support (Hg), and forced us to migrate to Git.